The current license is incompatible with the original library's license:
* This fork [added the GPL license on Oct 2, 2015](https://github.com/abkonsta/NLog.Extensions.AzureTableStorage/commits/master/LICENSE.txt)
* The original repository [added the MIT license on Jan 5, 2016](https://github.com/harouny/NLog.Extensions.AzureTableStorage/commits/master/License.txt)

As a derivative work, you must follow the intent of the original author whose work you extended, which means the license must be compatible with the original work's license. Unfortunately, the GPL license is a very restrictive license, whereas the MIT license is a very permissive license.

This PR changes the license from GPL to MIT, which ensures compatibility with the original work.

[Further reading](https://softwareengineering.stackexchange.com/questions/240612/forking-bsd-project-and-changing-license)